fix: failing stx mainnet tests
