+++ tests/mocks/constants.ts
@@ -5,9 +5,11 @@ export const TEST_ACCOUNT_1_TAPROOT_ADDRESS =
 
 // export const TEST_TESTNET_ACCOUNT_1_BTC_ADDRESS = 'tb1q3c7zyg58dd9hy07m77dv8es9vnpk8xad0yaw8y'
 export const TEST_ACCOUNT_1_STX_ADDRESS = 'SP297VG59W96DPGBT13SGD542QE1XS954X78Z75G0';
+// export const TEST_TESTNET_ACCOUNT_1_STX_ADDRESS = 'ST297VG59W96DPGBT13SGD542QE1XS954X4ZSWW2J';
 // export const TEST_ACCOUNT_1_PUBKEY =
 //   '02b6b0afe5f620bc8e532b640b148dd9dea0ed19d11f8ab420fcce488fe3974893';
-export const TEST_ACCOUNT_2_STX_ADDRESS = 'SPRE7HABZGQ204G3VQAKMDMVBBD8A8CYG6BQKHQ';
+// export const TEST_ACCOUNT_2_STX_ADDRESS = 'SPRE7HABZGQ204G3VQAKMDMVBBD8A8CYG6BQKHQ';
+export const TEST_TESTNET_ACCCOUNT_2_STX_ADDRESS = 'STRE7HABZGQ204G3VQAKMDMVBBD8A8CYKET9M0T';
 // export const TEST_ACCOUNT_2_BTC_ADDRESS = 'bc1qznkpz8fk07nmdhvr2k4nnea5n08tw6tk540snu';
 export const TEST_TESTNET_ACCOUNT_2_BTC_ADDRESS = 'tb1qkzvk9hr7uvas23hspvsgqfvyc8h4nngeqjqtnj';
